requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
openpyxl
streamlit-aggrid
kaleido
statsmodels
Kaggle
scikit-learn
pyarrow 
fastparquet
Seaborn
python-dotenv
reportlab
playwright
fpdf2
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4
